+++ .github/workflows/prove.yml
@@ -37,7 +37,6 @@ jobs:
         env:
           RPC_URL: ${{ secrets.RPC_URL }}
         run: |
-          echo $RPC_URL
           cd program
           cargo run -r
