+++ lib/src/provider/mod.rs
@@ -1,2 +1,3 @@
+pub mod cl_header;
 pub mod header;
 pub mod transaction;
